HappinessApp-Django/requirements.txt
argcomplete==1.10.0
asgiref==3.2.7
backports.zoneinfo==0.2.1
beautifulsoup4==4.8.0
boto3==1.14.20
botocore==1.17.20
certifi==2021.10.8
chardet==3.0.4
charset-normalizer==2.0.12
dj-database-url==0.5.0
Django==3.0.7
django-crispy-forms==1.9.1
django-heroku==0.3.1
django-storages==1.9.1
django-taggit==2.1.0
djangorestframework==3.13.1
docutils==0.15.2
docx2txt==0.8
EbookLib==0.17.1
exponent-server-sdk==2.0.0
extract-msg==0.23.1
gunicorn==20.0.4
idna==3.3
IMAPClient==2.1.0
jmespath==0.10.0
lxml==4.5.2
olefile==0.46
pdfminer.six==20181108
Pillow==9.1.0
psycopg2==2.9.3
psycopg2-binary==2.9.3
pycryptodome==3.9.8
python-dateutil==2.8.1
python-pptx==0.6.18
pytz==2020.1
requests==2.27.1
rollbar==0.16.2
s3transfer==0.3.3
six==1.12.0
sortedcontainers==2.2.2
soupsieve==2.0.1
SpeechRecognition==3.8.1
sqlparse==0.3.1
textract==1.6.3
tzlocal==1.5.1
urllib3==1.25.9
whitenoise==5.1.0
xlrd==1.2.0
XlsxWriter==1.3.3
